Check if there is a config file, else use default
