+++ src/Stryker.CLI/Stryker.CLI.UnitTest/Stryker.CLI.UnitTest.csproj
@@ -22,8 +22,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Include="stryker-config.json" CopyToOutputDirectory="Always" />
-    <None Include="filled-stryker-config.json" CopyToOutputDirectory="Always" />
+    <None Update="filled-stryker-config.json" CopyToOutputDirectory="Always" />
+    <None Update="stryker-config.json" CopyToOutputDirectory="Always" />
   </ItemGroup>
 
   <Target Name="PrintReferences" DependsOnTargets="ResolveProjectReferences;ResolveAssemblyReferences">
